% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjustments}
\alias{adjustments}
\title{Trade flows adjustments from MDB files}
\format{A data frame}
\usage{
adjustments
}
\description{
Trade flows adjustments from MDB files
}
\keyword{datasets}

