% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esdata2faoarea.R
\name{esdata2faoarea}
\alias{esdata2faoarea}
\title{Converts column `reporter` from Geonom codes to FAO area list codes.}
\usage{
esdata2faoarea(data, g2fmap)
}
\arguments{
\item{data}{Data frame with Eurostat trade data, containing 
column called `reporter`.}

\item{g2fmap}{Geonom to FAO area list mapping table: data frame 
containing columns geonom and fao.}
}
\value{
Data frame, similar to the original `data`, but 
  codes in `reporter` column converted.
}
\description{
Converts column `reporter` from Geonom codes to FAO area list codes.
}

