% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6faointerest}
\alias{hs6faointerest}
\title{6-digit HS codes what are of FAO's interest}
\format{A character vector of length 1061}
\usage{
hs6faointerest
}
\description{
Taken from `HS2012-6 digits Standard.xls` provided by Claudia and stored in `data-raw` folder of hsfclmap package.
}
\keyword{datasets}

