% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsfclmap}
\alias{hsfclmap}
\title{Most recent mapping between FCL and country specific HS}
\format{A data frame}
\usage{
hsfclmap
}
\description{
Most recent mapping between FCL and country specific HS
}
\keyword{datasets}

