% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadesdata.R
\name{loadesdata}
\alias{loadesdata}
\title{Function to load raw Eurostat trade data from csv file and procees it.}
\usage{
loadesdata(file = file.path(Sys.getenv("HOME"),
  "ce_combinednomenclature_unlogged_2014.csv.gz"))
}
\arguments{
\item{file}{Source csv file path.}
}
\description{
Function to load raw Eurostat trade data from csv file and procees it.
}
\details{
The function is copied and adapted from trade module. 
The function returns records sufficed the following
requirements: 
  * the chapters of interest, 
  * `stat_regime` #4, 
  * numeric values in reporter, partner and hs columns, 
  * length of hs codes is more than 2.
}

