% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadgeonom.R
\name{loadgeonom}
\alias{loadgeonom}
\title{Loads table for conversion of Geonom to FAO area list.}
\usage{
loadgeonom()
}
\value{
geonom2fao R-object stored in faoswsTrade package.
}
\description{
Loads table for conversion of Geonom to FAO area list.
}
\details{
The function created to avoid working with local files. 
It downloads the object from GitHub repository.
}

