% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtldata.R
\name{loadtldata}
\alias{loadtldata}
\title{Function to load raw Comtrade Tariffline data from csv file and procees it.}
\usage{
loadtldata(file = file.path(Sys.getenv("HOME"),
  "ct_tariffline_unlogged_2014.csv.gz"))
}
\arguments{
\item{file}{Source csv file path.}
}
\description{
Function to load raw Comtrade Tariffline data from csv file and procees it.
}
\details{
The function is copied and adapted from trade module. 
The function returns records sufficed the following
requirements: 
  * the chapters of interest, 
  * numeric values in reporter, partner and hs columns, 
  * length of hs codes is more than 2.
}

