% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtradefile.R
\name{loadtradefile}
\alias{loadtradefile}
\title{Function to load raw Comtrade or Eurostat data from csv file and procees it.}
\usage{
loadtradefile(file = file.path(Sys.getenv("HOME"),
  "ct_tariffline_unlogged_2014.csv.gz"), chapters = c(1:24, 33, 35, 38, 40:43,
  50:53), tariffline = grepl("tariffline", file),
  coltypes = ifelse(tariffline, "cci__i_c_______", "cc_ciic___"),
  colnames = if (tariffline) c("chapter", "reporter", "year", "flow", "hs")
  else c("chapter", "reporter", "hs", "flow", "stat_regime", "year"))
}
\arguments{
\item{file}{Source csv file path.}
}
\description{
Function to load raw Comtrade or Eurostat data from csv file and procees it.
}
\details{
The function is copied and adapted from trade module. 
The function returns records sufficed the following
requirements: 
  * the chapters of interest, 
  * numeric values in reporter and hs columns, 
  * length of hs codes is more than 6.
}

