% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualCorrections.R
\name{manualCorrections}
\alias{manualCorrections}
\title{Correction of mistakes in country-specific
HS->FCL mapping tables from MDB files}
\usage{
manualCorrections(map)
}
\description{
Correction of mistakes in country-specific
HS->FCL mapping tables from MDB files
}

