% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapupdate.R
\name{mapupdate}
\alias{mapupdate}
\title{Adds updates to HS->FCL map}
\usage{
mapupdate(maptable, update, startyear = NULL, endyear = 2050L)
}
\description{
Usually updates contain one HS code instead of HS range. The function duplicates HS column into `fromcode` and `tocode` columns and adds `rownumb` column based on the latest (the maximal) value available in the mapping table.
}

