% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade2014missing}
\alias{trade2014missing}
\title{Update of mapping data table for trade data 2014}
\format{A data frame with 563 rows and 4 columns
\describe{
  \item{area}{Integer. Reporter code.}
  \item{flow}{Integer. Trade direction.}
  \item{hs}{Character. HS commodity code.}
  \item{fcl}{Integer. FCL commodity code.}
}}
\usage{
trade2014missing
}
\description{
Update of mapping data table for trade data 2014
}
\keyword{datasets}

