% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6agri}
\alias{hs6agri}
\title{Another variant of HS codes of FAO's interest}
\format{A data frame with 42 rows and 2 columns
\describe{
  \item{FromCode}{Numeric. Lower limit of HS range.}
  \item{ToCode}{Numeric. Upper limit of HS range.}
}}
\usage{
hs6agri
}
\description{
Probably this one is better than previous `hs6faointerest` 
  as we got it after Carola discovered a mistake in old data set.
}
\keyword{datasets}

