% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsfclmap2}
\alias{hsfclmap2}
\title{Mappings between FCL and country specific HS starting from 2009.}
\format{A data frame}
\usage{
hsfclmap2
}
\description{
Mappings between FCL and country specific HS starting from 2009.
}
\keyword{datasets}

