% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsfclmap3}
\alias{hsfclmap3}
\title{HS->FCL country/year specific mapping table without duplicate links.}
\format{A data frame with 885013 rows and 6 columns
\describe{
  \item{startyear}{Integer. Year starting from which specified link is actual.}
  \item{area}{Integer. FAO area code.}
  \item{flow}{Interger. Import/Export}
  \item{fromcode}{Character. Starting HS code.}
  \item{tocode}{Character. Ending HS code.}
  \item{fcl}{Interger. FCL code.}
}}
\usage{
hsfclmap3
}
\description{
A dataset containing mapping table from Harmonized System country specific 
codes to FAO Commodity List codes. Also columns for dubugging of MDB-export process excluded.
}
\keyword{datasets}

